Fully merged SSA and SSAU
